refactor(crawler): streamline SHA1 handling and validation
- Removed redundant SHA1 handling code from the GCS client.
- Introduced a new `sha1` package to encapsulate SHA1 parsing logic.
- Enhanced the `FetchSHA1` method to utilize the new `sha1.Parse` function for improved clarity and maintainability.
- Updated dependencies in go.mod and go.sum for consistency.
