+++ pkg/crawler/gcs/client.go
@@ -1,7 +1,6 @@
 package gcs
 
 import (
-	"bytes"
 	"cmp"
 	"context"
 	"encoding/json"
@@ -12,7 +11,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/aquasecurity/trivy-java-db/pkg/index"
+	"github.com/aquasecurity/trivy-java-db/pkg/crawler/sha1"
 	"github.com/hashicorp/go-retryablehttp"
 	"github.com/samber/lo"
 	"golang.org/x/xerrors"
@@ -104,31 +103,7 @@ func (s *Client) FetchSHA1(ctx context.Context, itemName string) (string, error)
 	if err != nil {
 		return "", xerrors.Errorf("failed to get object %s: %w", itemName, err)
 	}
-	data = bytes.TrimSpace(data)
-
-	// Handle empty SHA1 files
-	// e.g.
-	//    https://repo.maven.apache.org/maven2/org/wso2/msf4j/msf4j-swagger/2.5.2/msf4j-swagger-2.5.2.jar.sha1
-	//    https://repo.maven.apache.org/maven2/org/wso2/carbon/analytics/org.wso2.carbon.permissions.rest.api/2.0.248/org.wso2.carbon.permissions.rest.api-2.0.248.jar.sha1
-	if len(data) == 0 {
-		return index.NotAvailable, nil
-	}
-
-	// Find a valid SHA1 hash in the content
-	parts := strings.Fields(string(data))
-
-	// Validate SHA1 as there are xxx.jar.sha1 files with additional data.
-	// e.g.
-	//   https://repo.maven.apache.org/maven2/aspectj/aspectjrt/1.5.2a/aspectjrt-1.5.2a.jar.sha1
-	//   https://repo.maven.apache.org/maven2/xerces/xercesImpl/2.9.0/xercesImpl-2.9.0.jar.sha1
-	for _, part := range parts {
-		if len(part) == 40 && isHexString(part) {
-			return part, nil
-		}
-	}
-
-	// Record wrong SHA1 digests so we can skip them in the future
-	return index.NotAvailable, nil
+	return sha1.Parse(data), nil
 }
 
 // Internal methods
@@ -283,14 +258,4 @@ func (s *Client) listObjects(ctx context.Context, params RequestParams, extracto
 			pageToken = result.NextPageToken
 		}
 	}
-}
-
-// isHexString checks if a string contains only hex characters
-func isHexString(s string) bool {
-	for _, c := range s {
-		if !((c >= '0' && c <= '9') || (c >= 'a' && c <= 'f') || (c >= 'A' && c <= 'F')) {
-			return false
-		}
-	}
-	return true
 }
